refactor middleware into other files
